Move KubeCluster init kwarg tests to test_unit/test_kubernetes.py
rm duplicate bad args test


rm duplicate init args test
